+++ src/app/admin-workexperience/admin-workexperience.component.ts
@@ -15,6 +15,8 @@ export class AdminWorkexperienceComponent {
   workExperience: WorkExperience[] = [];
   myWorkExperience: WorkExperience = new WorkExperience();
   selectedId?: string = '';
+  errorMessage: string = '';
+
 
   constructor(public workExperienceService: WorkExperienceService) {
     console.log(this.workExperienceService);
@@ -30,12 +32,25 @@ export class AdminWorkexperienceComponent {
     });
   }
 
+  isValid(): boolean {
+    const { company, position, startDate, endDate, location, accomplishment } = this.myWorkExperience;
+    return !!(company?.trim() && position?.trim() && startDate?.trim() && endDate?.trim() && location?.trim() && accomplishment?.trim());
+  }
+
   AgregarJob() {
+   if (!this.isValid()) {
+      this.errorMessage = 'Please fill out all fields.';
+      return;
+    }
+
+    this.errorMessage = '';
+if (confirm('Are you sure you want to update this job?')) {
     if (this.selectedId) {
       this.workExperienceService.updateWorkExperience(this.selectedId, this.myWorkExperience).then(() => {
         console.log('Updated work experience successfully!');
         this.resetForm();
       });
+    }
     } else {
       this.workExperienceService.createWorkExperience(this.myWorkExperience).then(() => {
         console.log('Created new work experience successfully!');
@@ -45,10 +60,12 @@ export class AdminWorkexperienceComponent {
   }
 
   deleteJob(id?: string) {
+   if (confirm('Are you sure you want to delete this job?')) {
     this.workExperienceService.deleteWorkExperience(id).then(() => {
       console.log('Deleted work experience successfully!');
     });
     console.log(id);
+  }
   }
 
   updateJob(id?: string) {
@@ -56,7 +73,7 @@ export class AdminWorkexperienceComponent {
     if (jobToEdit) {
       this.myWorkExperience = { ...jobToEdit };
       this.selectedId = id;
-      this.btnTxt = "Confirmar actualización";
+      this.btnTxt = "Save";
     }
   }
 
@@ -64,6 +81,7 @@ export class AdminWorkexperienceComponent {
     this.myWorkExperience = new WorkExperience();
     this.selectedId = '';
     this.btnTxt = "Agregar";
+    this.errorMessage = ''; //reset el mensaje
   }
 }
